02/solutions/03: Fix name to avoid clash with 4th exercise
The 2.3 and 2.4 exercise are the same, with the difference that one is
implemented with structs and the other with enums.

The initial setup for 2.3 used the name of the exercise without
containing the differentiation.

This commit rename the directory, crate and README title with the
differentiation to avoid clashes and keep it homogeneous with the name
that will give to 2.4.